requirements.txt
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.1
django-appconf==1.0.2
django-imagekit==4.0.1
gunicorn==19.6.0
idna==2.5
olefile==0.44
pilkit==2.0
Pillow==4.2.1
psycopg2==2.6.2
python-dateutil==2.6.1
pytz==2017.2
pyuploadcare==2.2.1
requests==2.18.1
six==1.10.0
urllib3==1.21.1
whitenoise==3.2
